requirements.txt
langchain == 0.2.14
langchain_community == 0.2.12
langchain_core == 0.2.34
langchain_openai == 0.1.22
langchain_anthropic == 0.1.23
langchain_mistralai == 0.1.12
langchain_groq == 0.1.9
python-dotenv==1.0.1
streamlit == 1.31.1